# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.ProductCategory:
    enum:
    - Grocery
    - FruitAndVeg
    - Meat
    - Poultry
    - Seafood
    - Bakery
    - Liquor
    - GeneralStore
    type: string
    x-enum-varnames:
    - ProductCategoryGrocery
    - ProductCategoryFruitAndVeg
    - ProductCategoryMeat
    - ProductCategoryPoultry
    - ProductCategorySeafood
    - ProductCategoryBakery
    - ProductCategoryLiquor
    - ProductCategoryGeneralStore
...
  enums.FeeType:
    enum:
    - PERCENTAGE
    - FIXED_AMOUNT
    type: string
    x-enum-varnames:
    - Percentage
    - FixedAmount
...
  request.CreateLateFeeRuleRequest:
    properties:
      days_overdue_max:
        type: integer
      days_overdue_min:
        minimum: 1
        type: integer
      establishment_id:
        description: Optional, if rule is for specific establishment
        type: integer
      fee_type:
        $ref: '#/definitions/enums.FeeType'
      fee_value:
        type: number
      name:
        type: string
    required:
    - days_overdue_max
    - days_overdue_min
    - fee_type
    - fee_value
    - name
    type: object
  request.CreateProductRequest:
    properties:
      category:
        $ref: '#/definitions/entities.ProductCategory'
      description:
        type: string
      establishment_id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category
    - description
    - establishment_id
    - name
    - price
    - stock
    type: object
...
  request.UpdateLateFeeRuleRequest:
    properties:
      days_overdue_max:
        type: integer
      days_overdue_min:
        minimum: 1
        type: integer
      fee_type:
        $ref: '#/definitions/enums.FeeType'
      fee_value:
        type: number
      name:
        type: string
    type: object
  request.UpdateProductRequest:
    properties:
      category:
        $ref: '#/definitions/entities.ProductCategory'
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category
    - description
    - name
    - price
    - stock
    type: object
...
  response.AdminResponse:
    properties:
      created_at:
        type: string
      establishment:
        $ref: '#/definitions/response.EstablishmentResponse'
      establishment_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.EstablishmentResponse:
    properties:
      address:
        type: string
      admin:
        $ref: '#/definitions/response.AdminResponse'
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      products:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      ruc:
        type: string
      updated_at:
        type: string
    type: object
...
  response.LateFeeRuleResponse:
    properties:
      days_overdue_max:
        type: integer
      days_overdue_min:
        type: integer
      establishment_id:
        type: integer
      fee_type:
        $ref: '#/definitions/enums.FeeType'
      fee_value:
        type: number
      id:
        type: integer
      name:
        type: string
    type: object
  response.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/entities.ProductCategory'
      created_at:
        type: string
      description:
        type: string
      establishment:
        $ref: '#/definitions/response.EstablishmentResponse'
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing finances in small businesses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Final Assignment Finance API Rest
  version: "1.0"
paths:
...
  /api/v1/establishments/{establishment_id}/products:
    get:
      description: Retrieve products associated with a specific establishment ID.
      parameters:
      - description: Establishment ID
        in: path
        name: establishment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get products by establishment ID
      tags:
      - Products
...
  /api/v1/establishments/{establishmentID}/late-fee-rules:
    get:
      description: Retrieves all late fee rules for a specific establishment.
      parameters:
      - description: Establishment ID
        in: path
        name: establishmentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.LateFeeRuleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get late fee rules by establishment ID
      tags:
      - LateFeeRules
...
  /api/v1/establishments/{id}/products:
    post:
      consumes:
      - application/json
      description: Associate a list of products with an establishment.
      parameters:
      - description: Establishment ID
        in: path
        name: id
        required: true
        type: integer
      - description: List of product IDs
        in: body
        name: products
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register products for an establishment
      tags:
      - Establishments
...
  /api/v1/late-fee-rules:
    get:
      description: Retrieves all late fee rules.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.LateFeeRuleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all late fee rules
      tags:
      - LateFeeRules
    post:
      consumes:
      - application/json
      description: Creates a new late fee rule for an establishment (or globally).
      parameters:
      - description: Late fee rule details
        in: body
        name: lateFeeRule
        required: true
        schema:
          $ref: '#/definitions/request.CreateLateFeeRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.LateFeeRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a late fee rule
      tags:
      - LateFeeRules
  /api/v1/late-fee-rules/{id}:
    delete:
      description: Deletes a late fee rule by its ID.
      parameters:
      - description: Late Fee Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a late fee rule
      tags:
      - LateFeeRules
    get:
      description: Retrieves a late fee rule by its ID.
      parameters:
      - description: Late Fee Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LateFeeRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get late fee rule by ID
      tags:
      - LateFeeRules
    put:
      consumes:
      - application/json
      description: Updates an existing late fee rule by its ID.
      parameters:
      - description: Late Fee Rule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated late fee rule details
        in: body
        name: lateFeeRule
        required: true
        schema:
          $ref: '#/definitions/request.UpdateLateFeeRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LateFeeRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a late fee rule
      tags:
      - LateFeeRules
...
  /api/v1/products:
    get:
      description: Retrieve a list of all products.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ProductResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product associated with an establishment.
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new product
      tags:
      - Products
  /api/v1/products/{id}:
    delete:
      description: Delete a product by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a product
      tags:
      - Products
    get:
      description: Retrieve a product by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing product
      tags:
      - Products
...
swagger: "2.0"